feat: review-less protection.
logion-network/logion-internal#1123
