+++ src/legal-officer/LegalOfficerPaths.tsx
@@ -10,9 +10,6 @@ import {
 
 export const HOME_PATH = LEGAL_OFFICER_PATH;
 
-export const PROTECTION_REQUESTS_RELATIVE_PATH = '/protection';
-export const PROTECTION_REQUESTS_PATH = LEGAL_OFFICER_PATH + PROTECTION_REQUESTS_RELATIVE_PATH;
-
 export const RECOVERY_REQUESTS_RELATIVE_PATH = '/recovery';
 export const RECOVERY_REQUESTS_PATH = LEGAL_OFFICER_PATH + RECOVERY_REQUESTS_RELATIVE_PATH;
