--- src/legal-officer/ProtectionRequests.css
@@ -1,3 +1,0 @@
-.ProtectionRequests .Tabs {
-    margin-bottom: 50px;
-}
