--- src/legal-officer/ProtectionRequestsHistory.test.tsx
@@ -1,19 +1,0 @@
-jest.mock('./LegalOfficerContext');
-jest.mock('../logion-chain');
-jest.mock('./Model');
-
-import { shallowRender } from '../tests';
-import ProtectionRequestsHistory from './ProtectionRequestsHistory';
-import { setProtectionRequestsHistory } from './__mocks__/LegalOfficerContextMock';
-import { PROTECTION_REQUESTS_HISTORY } from './TestData';
-
-test("Renders null with no data", () => {
-    const tree = shallowRender(<ProtectionRequestsHistory />);
-    expect(tree).toMatchSnapshot();
-});
-
-test("Renders requests history", () => {
-    setProtectionRequestsHistory(PROTECTION_REQUESTS_HISTORY);
-    const tree = shallowRender(<ProtectionRequestsHistory />);
-    expect(tree).toMatchSnapshot();
-});
