+++ src/loc/issuer/IssuerSelectionFrame.tsx
@@ -20,7 +20,6 @@ export default function IssuerSelectionFrame() {
         if (locState !== null && (locState instanceof OpenLoc || locState instanceof ClosedCollectionLoc)) {
             (async function() {
                 const issuers = await locState.legalOfficer.getVerifiedIssuers();
-                console.log(issuers)
                 setIssuerSelections(issuers);
             })();
         }
