--- src/wallet-user/trust-protection/ProtectionRefusal.css
@@ -1,19 +1,0 @@
-.ProtectionRefusal .Title {
-    text-align: center;
-    margin: 20px;
-}
-
-.ProtectionRefusal .Button {
-    margin: 15px;
-}
-
-.ProtectionRefusal .select,
-.ProtectionRefusal .restart {
-    display: flex;
-    width: 100%;
-    flex-wrap: wrap;
-}
-
-.ProtectionRefusal .ButtonGroup {
-    width: auto;
-}
