+++ src/wallet-user/trust-protection/TrustProtection.tsx
@@ -1,9 +1,6 @@
 import {
-    AcceptedProtection,
     NoProtection,
-    PendingProtection,
     UnavailableProtection,
-    RejectedProtection
 } from "@logion/client";
 
 import { useUserContext } from "../UserContext";
@@ -43,12 +40,6 @@ export default function TrustProtection() {
             return <GoToRecovery />;
         } else if(protectionState.protectionParameters.isActive) {
             return <ProtectionRecoveryRequest type='activated' />;
-        } else if(protectionState instanceof PendingProtection) {
-            return <ProtectionRecoveryRequest type='pending' />;
-        } else if(protectionState instanceof AcceptedProtection) {
-            return <ProtectionRecoveryRequest type='accepted' />;
-        } else if (protectionState instanceof RejectedProtection) {
-            return <ProtectionRecoveryRequest type='rejected' />;
         } else {
             return null;
         }
